Honor mgmt_ip boolean in FCR-OSS
Summary: Looks like mgmt_ip flag was not being read on the server. I'm guessing this is a bug. One other possibility is that this wasn't done intentionally since this code is going to be open-sourced. Since sursingh is off-the-grid, I'll leave it for starskywongfb / mlal to decide whether this diff should go in or not.

Reviewed By: starskywongfb

Differential Revision: D5601351

fbshipit-source-id: a9d0612a856a2ccdcbb808cb691c95b5833c8f21