save groupdict in the ResponseMatch
Summary:
We use regex to match prompt as we receive data. The data is stored in
bytebuffer and is manipulated in place. After we read the data the buffer may
get overwritten. The match object seems to be directly referring the data in
bytebuffer. This causes a problem when we try to find the matched groups.

Now we will save the matched groups before reading the data.

Reviewed By: starskywongfb

Differential Revision: D5732849

fbshipit-source-id: ba83168a09043d18c47541cd1e6da078f6da81db