Inject a newline into recv buffer to simplify prompt processing
Summary:
Devices sometimes don't send newline with first prompt. This causes issues with
prompt processing. Forcibly sending a newline address this but causes other
issues as we sometime get out of sync with the device.

Now we will inject a "newline" when a stream is first created. This ensures
that we always get a newline. If the device does send a newline, then this will
be a nop and will be ignored.

Also added a local only integration test to for validating the commands. Will
later try to add this to our service foundary canary

Differential Revision: D5903514

fbshipit-source-id: c2c44235187de905e3b0976265cc5cf78a89e9f6