Don't use radix tree for console connections
Summary:
We use hostnames to connect to the console servers. Radix tree doesn't work for
hostnames. Ideally we should lookup the console server hostnames and then
select the proxy based on that. That will take sometime to implement. In the
interim, we will not use the radix tree for deciding proxy for console servers.
We use proxy for console servers based on their names

Reviewed By: starskywongfb

Differential Revision: D6991742

fbshipit-source-id: 68119385045c1b9a4b0bab30800fc2ebc945ba92