Add netconf support in FCR
Summary:
Initial version of Netconf support for FCR. We don't try to do any fancy processing for NETCONF. Both the command and reply are opaque to FCR.

FCR will send the initial Hello Message to the peer.

The Hello message from the peer will be forwarded to client in the first reply. This is mainly useful for interactive sessions, where the client may want to look at the capabilities of the peer before sending the commands

Reviewed By: starskywongfb

Differential Revision: D7801218

fbshipit-source-id: 3b05de5c13c166ea1afa31ecc97022ae44d2ab95