add --optical_node to argument mode to support tl1
Summary: TL1 command is optical node-oriented. This diff is to add a new command line option '--optical_node' for the users to specify the regex of the optical node names they want to run the TL1 command on. After this diff, user can use the argument mode to run TL1 command with FCR. Device session mode is not support yet but in my next diff.

Differential Revision: D7958739

fbshipit-source-id: d4bf2b2243c65434b720ce7524abf1435c0a7163