Adding request uuid and session id to FCR run command scuba table
Summary:
Add thrift request uuid and session id in the FCR per command scuba table (https://fburl.com/scuba/wuj5teaj) to map entries in the this table to those in the FCR thrift usage  table (https://fburl.com/scuba/pkh6pn3l).

To implement this, we added an additional argument uuid to the FCR thrift APIs. The FCR thrift handler at the server side will append the uuid of each thrift call along the call sequence to this argument

We also append the uuid in the exception message for debugging purpose

Differential Revision: D8037752

fbshipit-source-id: cd88f08d5671639619ddf4d6ac4c2be6bbf8eeed