log exception in session creation and closing
Summary:
In FCR we only log the result and exception of the _run_command method. But the connection session is created before calling this method. If there's an exception in session creation, the program will be terminated without calling this method, and hence will not be logged.

This diff logs the exception in session creation and closing to scuba as well

Differential Revision: D13606727

fbshipit-source-id: 162e78404f468a768c7452d35f02ba44ccf065a9