Enable new NAT/amphibolite
Summary:
This schedules a separate task to handle the lease acquiry & expiration, signaling with an async event. We can get all we need from configerator & fbwhoami. This is hitting the new, more performant control plane.

If the control plane is down when our task has restarted, check if the dataplane is okay by pinging our v4 DNS. Success means a valid lease still exists for this host.

Differential Revision: D14072373

fbshipit-source-id: eedbf9ec26a847712ebe0e6792407d10a6ed93ed