Configlets_push load balancing
Summary: In bulk_run, `_bulk_session_limit` was given a setter so that the class variable can be changed as opposed to an attribute of the class. It currently makes no functional difference in the code since we only run one command handler per instance of FCR, but it appears that this was the original intent of the author. It also makes more sense, since the resources used on a session belong to an instance of FCR, not to a command handler class (i.e. if you create a new `CommandHandler`, you don't gain the compute, memory, or network resources run more sessions concurrently. Thus it should be based on each instance of FCR)

Differential Revision: D16116334

fbshipit-source-id: 867cf16dc288d8e3dfaa5dc9a6eadf145637bea6