Prevent FCR to start when device_db fails to fetch device info for the frist time
Summary:
Currently FCR does not have preventions when device_db fails to fetch device data, thus it continues to setup the server regardless. We do not want this to happen so adding a mechanism which would led device_db to timeout if it fails to fetch device info for the first time.

Added few unittests accordingly

Differential Revision: D18308675

fbshipit-source-id: 7e12b3f832e048344002a1f2fbf2dc2317dc94c6