Apply µsort to fbcode/nettools
Summary:
Runs pyfmt (with µsort instead of isort) over a subset of opted-in
Python sources in fbsource. This is a new, safer import sorting
system. See https://usort.readthedocs.io/ for details.

Of note is that µsort ignores `# isort:skip_file` directives, but
it should not break common patterns like changes to `sys.path` or
function calls interleaved with import statements. There is also
a configured list of modules with known import-time side effects
that µsort will always refuse to reorder. In these cases, µsort
will split the file into logical "blocks" of imports, and only
sort imports within each separate block.

We recognize this is a potentially dangerous change, especially
for modules that contain isort skip directives, but we believe this
will provide a better, safer experience. If there are any changes
in this file that you believe are unsafe, please request changes
and let us know where and why. We will do our best to accomodate.

Reviewed By: zertosh

Differential Revision: D24775445

fbshipit-source-id: f66e6ddb7d52d93af4beda8cefb1768caae60114