refactor console session build_prompt_re() as a setter
Summary: build_prompt_re was expected to be used as a setter function of the global prompt regex of the console session. However, the logic of the code was actually only returning the current prompt regex (which is hardcoded), effectively adding no functionality. Now the function actually sets the global prompt regex.

Differential Revision: D26609203

fbshipit-source-id: b8a67851afd3819bee44f926be3a7b036c6ed354