Avoid returning NoneType ip address from DeviceInfo
Summary:
# Summary
This diff is first diff to improve the IP fallback logic in FCR. When FCR decides on which IP to connect to, it keeps a preference list (IPv6 over IPv4, etc) and choose the first IP that is pingable. When the IPs in the preference list are all not pingable, FCR fallback to the "ip" field in the corresponding "ip" field of the Skynet `NetworkDevice` obj.
The issue here is that, the "ip" field of a `NetworkDevice` can be `None` and FCR would fail to connect to the IP obviously, and user will see an exception such as:
```
Traceback (most recent call last):
  File "fbnet/command_runner/command_session.py", line 263, in __aenter__
    await self.setup()
  File "fbnet/command_runner/command_session.py", line 147, in wrapper
    return await fn(self, *args, **kwargs)
  File "fbnet/command_runner/command_session.py", line 347, in setup
    self._create_connection(), self.open_timeout, loop=self._loop
  File "/usr/local/fbcode/platform007/lib/python3.7/asyncio/tasks.py", line 442, in wait_for
    return fut.result()
  File "fbnet/command_runner/command_session.py", line 648, in _create_connection
    await super()._create_connection()
  File "fbnet/command_runner/command_session.py", line 340, in _create_connection
    await self.connect()
  File "fbnet/command_runner/command_session.py", line 373, in connect
    raise e
  File "fbnet/command_runner/command_session.py", line 367, in connect
    self._cmd_stream = await self._connect()
  File "fbnet/command_runner/command_session.py", line 854, in _connect
    elif self._devinfo.should_nat(ip):
  File "fbnet_data.py", line 82, in should_nat
    return should_nat_impl(ip, self.service.ENABLE_NAT)
  File "neteng/statelessnat/client/nat.py", line 149, in should_nat
    return enable_nat and ip_address(ip).version == 4
  File "ipaddress.py", line 168, in ip_address
    address)
ValueError: None does not appear to be an IPv4 or IPv6 address
```
 ---
In this diff, added additional logic to fallback to the first not None IP address instead, when None of the preferred IPs is Pingable. Raise an exception if we cannot find any useable IP for the device
 ---
Next diff:
- Improve exception message upon connection failure when the ip that is used to connect is not pingable

Reviewed By: hegara

Differential Revision: D26475389

fbshipit-source-id: b2f25846c504d58e34f56753f95828637f79c34a