Consolidate IGNORE_CONSOLE and CONSOLE_PROMPTS
Summary: We greatly simplify the logic of building the console prompts dictionary by consolidating the IGNORE_CONSOLE and CONSOLE_PROMPTS dictionaries. This is part of an effort to make the console session command running logic simpler and more robust. This diff also allows us to omit the field entirely, in the case there were no desired "ignore" prompts. In the previous logic, there was no way to achieve this (unless you wanted to declare an impossible regex like $a which is very computationally expensive to match against).

Differential Revision: D26751523

fbshipit-source-id: 1836bf2be28f76dccca848f56979f2790801bec8