Implement new interactive prompts handling logic
Summary:
We implement new logic for handling new interactive prompts if they exist

This is the last diff before things can really start to break.

This is the last diff that refers to the old cconf file: https://fburl.com/diffusion/7hf9tc36

There will be a new cconf file generated in D26824323.

Deployment plan is as follows:
1. Deploy stack of diffs [5/N](D26823037) -> [8/N](D26823037) and wait for them to hit production.
2. Land new cconf D26824323
3. Deploy [10/N](D26823460) which hooks-up to the new cconf in D26824323
4. Wait for [10/N](D26823460) to hit production completely and make sure no instances are referencing https://fburl.com/diffusion/0we3w6j4 anymore.
5. Optional: deprecate (or delete) https://fburl.com/diffusion/0we3w6j4
6. Enjoy

Differential Revision: D26823404

fbshipit-source-id: 950aeb761fd5a1a6d8a22446bcbbc083e7475a61