Remove blocking call when querying fburl service
Summary:
As a followup item for the sev S226358, we found out that there are many blocking calls / processing in FCR that may block the event loop

 ---

One blocking IO we found out was when creating error messages. Each error message will call the `fburl` method from ScubaURL, this method calls the external fburl server with blocking calls: https://fburl.com/diffusion/wx8wsqef

And FCR does that two times per failed request. This can be big burden on the event loop

So in this diff, added a util function called `create_fburl` that queries the fburl service in an async manner, to replace the blocking call

Differential Revision: D27418050

fbshipit-source-id: 70c48ec76db3cb26cc066322d18f9f15e674f5c8