Fix debug scuba URL by retrieving uuid from GlobalNamespace
Summary:
**What**
When a request throws an exception, we return the stacktrace and along with two fburl's for the scuba query where they can see the FCR thrift usage logs and FBnet Command logs to help users self-troubleshoot.
 ---
**How**
We leverage the ```GlobalNamespace``` module which stores context within the thread to retrieve the request uuid, eliminating the need for retrieving the uuid from the kwargs which was messed up by the ```ensure_uuid``` decorator.
 ---

Differential Revision: D28367058

fbshipit-source-id: 4d5e1347225b5e9bc383da9f7f4289f815938212