Implement support for custom clear command per request
Summary:
## Why
From user: "fcr appends \x15 character initially which is ascii equivalent of ctrl+U. What it does is it clears the cli so that the cli is clean before running any command. For the ciena 6500 chassis, when we run commands via fcr in config mode, it isn't able to recognize \x15 and thats why any command that we run fails (connection is successful but the command fails". config mode is port 28888. Normal ssh (default is port 20002 runs fine. It knows what \x15 is)
## How
1. Add new optional field in the Device struct. The field is a string but expects in bytes as it is later encoded on server-side. This makes the field configurable by users during each request. See {D29019874}
2. At the point where the "precmd" (the clearing command) is set, either choose the specified "clear_command" if it is set, and encode it OR fall-back to the default clear_command (if not set by user's request). Currently the default is "\x16".
3. Updated unit tests.

Differential Revision: D29019142

fbshipit-source-id: 3eb3a0b7901404d9a563f39bf3bd06e520222c80