Add IP failover option to Device struct
Summary:
Adding option to failover to other IPs of the device if the connection with the IP chosen by FCR does not succeed.

Especially needed for FNA devices where sometimes ssh to the IPv6 address does not work but IPv4 does. In such cases, FCR takes the first IP (here IPv6), fails on it and returns with an error. Failing over to IPv4 would help in such cases.

An optional field to the device struct can help users choose whether they want to opt for a failover or error out if the first IP/IP they provide does not connect

Differential Revision: D31405672

fbshipit-source-id: a68d7cd462552de5114e9005e412368b820cc51b