Remove DNS dependency on Console Session
Summary:
Background and motivation can be found in https://fb.workplace.com/groups/772600260163354/permalink/1091770751579635/
 ---
The goal is to remove dependency on DNS for FCR for console sever connections.

Historically, console server is not modeled in FBNET, but this is not the case anymore. So this dependency on DNS can be easily avoided by fetching the console server information inside our FBNET cache in FCR and this diff's goal is that.

Reviewed By: ItayTamary

Differential Revision: D32444043

fbshipit-source-id: be09dc51a377aae4e1c2b3d60592b1d1e653cf0b