Fallback to other ips if command fails to run on device
Summary:
For devices such as FNAs which are not covered by Netsonar and considered not Pingable, the first IP of the preferred IP lists from FBNet is selected to make the connection.

It is possible that the connection fails for many reasons. In that scenario, it is expected to have failovers to try connecting through the remaining IPs.

Differential Revision: D31353724

fbshipit-source-id: 2ffd650262ab83aa088eab73ce885f9111ed13f7