Large Command Execution Performance Improvement
Summary:
NOTE: This diff's change is identical to D27897576. Given that D27897576 is very old, I decided to just open a new diff to avoid messy merge conflicts

# Summary

In https://fb.workplace.com/groups/772600260163354/permalink/963513437738701/, user reported that the run api is running slow. The reason behind this long response time is because internally FCR would split the command by line and send the split command one by one and perform regex matching on the stream per split command. For example, if user sends FCR the below command:

```
show version\nshow version
```
FCR will send the first `show version` to the device, then performs regex matching, then sends the other one and performs regex matching. In pseudo code it looks like:
```
cmds = "show version\nshow version".splitline()
for cmd in cmds:
    send(cmd)
    result = regex_match(buffer)
    output.append(result)
return output
```

# Change of this diff
Alternatively, one way to improve this behavior is to send everything at once to the device, and match whatever is available in the stream. This will reduce the # of times FCR needs to write to the stream from linear to constant, also while FCR is matching the regex in the stream, device will process the commands at the same time.

The approach we are taking in this diff is to keep a secondary buffer (bytearray) that will store each line we found in the stream, and match the secondary buffer instead of the primary buffer. This will ensure that FCR won't overshoot with the regex matching when the primary buffer may contain multiple device prompt.

An option `split_single_command_by_line` is added to let user define the desired behavior. By default, this feature will be disabled for all requests to ensure backward compatibility. Only request that has `split_single_command_by_line = "no"` will enable this feature

Reviewed By: ItayTamary

Differential Revision: D32957014

fbshipit-source-id: d5349007f50d8f21baca6cc1d3f7f0b119fc9c97