Raise DisconnectError with correct arguments
Summary:
When credentials are incorrect for failover mode, the retry mechanism adds custom logic to the exception with the IPs that have been tried.

Of the list of known exceptions https://fburl.com/code/01qselgt, asyncssh.misc.DisconnectError requires a reason attribute. As the retry mechanism did not consider this, it errored out with a TypeError indicating a argument `reason` was missing.

Adding that condition in this diff

Differential Revision: D33947911

fbshipit-source-id: 27d418893a2433e299d5acbfd65b47cf40125fb4