FBNet Vendor.name deprecation: Phase 1
Summary:
## TL;DR
This diff should remove all usage of vendor.name as part of the deprecation of the field in FBNet.

## What
`Vendor.name` is the most important field in the functioning of FCR.

FCR's core functionality almost sole relies on `Vendor.name` to determine how it will interact with a device. The field signals to FCR which session to use to connect with the device and what prompts to match against. Because of this, the usage of the field is wide-spread throughout the production code and the unit-tests are written such that they validate explicit referencing of the field (making them brittle to this deprecation).

We risk breaking the core functionality of FCR and therefore significant automation of our network.

## How

Vendor.name is a non-standardized free-form field representing a pseudo-vendor of a device. E.g. IOS and NXOS can both be values in the vendor.name even though both are actually different operating systems of the vendor Cisco meaning that `Vendor.name` is misused.

Free-form string -> Enum: Reduces human-error and bugs

Alternatively, we can either use `HardwareVendor` or `OSType`. We have chosen `OSType` as this provides us with more granularity in controlling how FCR interacts with devices.

Specifically the new field is `desired_platform.os_type`

https://www.internalfb.com/code/fbsource/[3ecab25fd39c]/fbcode/configerator/structs/nettools/hardware/hardware.thrift?lines=109

There is only one special case that needs to be handled and that is of `Ciena 6500` optical devices. Both regular Ciena devices and 6500 devices are of the same HardwareVendor and OSType making them difficult to systematically differentiate through the traditional mechanism of an FBNet field. Therefore, we hard-code  FCR to check that if a device is a 6500 chassis model, to use the prompt configuration of `6500` over Ciena. Specifically, Ciena 6500 models need FCR to connect via their 20002 port.

## Compromises
FCR is written such that it assumes all behavior with a device is determined by the device's vendor. Once we started using multiple platforms from a single vendor in our network, instead of changing the naming convention in the codebase to reflect behavior by OSType, we doubled-down and begin misusing vendor as a reference to OSType which is where our tech-debt began to build. As such, the referencing of vendor in variable names and classes is wide-spread and can not be easily fixed by a code-mod (e.g. `FBDeviceVendor` and `vendor_name` in thrift structs) and verifying correctness of behavior as to not break FCR is a manual process given that we won't catch any incorrect use of a vendor variable at compilation time because Python is dynamically typed. This naming convention in the codebase would be worth doing (2 weeks) and improve code quality significantly if we weren't currently planning on rewriting FCR from scratch anyways, however, the cost/benefit does not make sense at this time. Therefore, variables such as `vendor_name` and `.name` will still be used throughout the codebase. The scope of this diff is to only stop querying for and referencing the FBNet field of `Vendor.name` as to not block progress on the effort to deprecate the field in FBNet

# Deployment Plan
FCR's logic is determined in both FBCode and CFGR and so changes have to be made across repos. Additionally, we need to do a calculated multi-step deployment so that we can roll-back quickly and reduce the risk of pro-longed regression of the FCR service.

### Phase 1: Initial Deployment
1. CFGR: Land new transitional file {D30766584} that uses OSType instead of `vendor.name` [ItayTamary]
2. FBCode: Land this diff {D35327591} where FCR consumes the "transitional" configs that use OSType [ItayTamary]

During deployment some FCR instances will reference the old CFGR file and use `vendor.name` while recently updated instances will reference the new transitional FCR file and use `OSType`.

If we detect a regression in updated instances, we can quickly roll-back to the previous FBPKG and eliminate any disruption of service to our users.

Once this diff is fully-rolled out to all instances and we determine that the
change is safe with no regression we can begin to prepare for clean-up.

## Phase 2: Preparing for clean-up
1. CFGR: Overwrite the contents of the transitional file to the regular file. E.g. `device_vendors_transitional.thrift` => `device_vendors.thrift` and land [youwang]
2. FBCode: Have all consumed configs point to the non-transitional files [youwang]

During deployment some instances will consume `_transitional` configs while the recently updated instances will consume the non-transitional configs.

Once fully deployed, all instances will consume the non-transitional we can complete the clean-up.

## Phase 3: Clean-Up
1. CFGR: Delete all transitional files as they are no longer being consumed by the FCR service. [youwang]

## Win

- Contribute to the removal of tech-debt of a redundant FBNet field
- Usage of a standardized enum field instead of a free-form string making FCR behavior more predictable and reduces chances of human error when configuring how FCR will interact with a device.

Differential Revision: D35327591

fbshipit-source-id: 54325e6befceda8dcf4b053d605f0af8595f4ed6